Update Components Name
