+++ packages/shared/src/components/Provider/index.tsx
@@ -1,0 +1,36 @@
+import { ThemeProvider } from '@emotion/react';
+import { QueryClient, QueryClientProvider } from '@tanstack/react-query';
+import { ReactQueryDevtools } from '@tanstack/react-query-devtools';
+import { useState } from 'react';
+import { GlobalStyle, theme } from 'shared/styles';
+
+interface Props {
+  readonly children: React.ReactNode;
+}
+
+const Provider: React.FC<Props> = ({ children }) => {
+  const [queryClient] = useState(
+    () =>
+      new QueryClient({
+        defaultOptions: {
+          queries: {
+            retry: false,
+            staleTime: 300000, // 5분
+            gcTime: 300000,
+          },
+        },
+      })
+  );
+
+  return (
+    <ThemeProvider theme={theme}>
+      <QueryClientProvider client={queryClient}>
+        <ReactQueryDevtools />
+        <GlobalStyle />
+        {children}
+      </QueryClientProvider>
+    </ThemeProvider>
+  );
+};
+
+export default Provider;
