--- packages/shared/src/components/Wrapper/index.tsx
@@ -1,17 +1,0 @@
-import { ThemeProvider } from '@emotion/react';
-import { GlobalStyle, theme } from 'shared/styles';
-
-interface Props {
-  readonly children: React.ReactNode;
-}
-
-const Wrapper: React.FC<Props> = ({ children }) => {
-  return (
-    <ThemeProvider theme={theme}>
-      <GlobalStyle />
-      {children}
-    </ThemeProvider>
-  );
-};
-
-export default Wrapper;
