+++ packages/shared/src/components/index.ts
@@ -1,2 +1,2 @@
-export { default as Wrapper } from './Wrapper';
+export { default as Provider } from './Provider';
 export { default as Header } from './Header';
